opsui/traits: relocate LightCell/Card into traits/lighting folder and update component imports
Jira: N/A
