+++ ui/conductor/src/traits/lighting/LightCell.vue
